Refactor to move the SSE task and message processing to a dedicated service
